Restructure project layout
Instead of a single binary, this commit creates a separate binary
for each person, with shared functionality (server communication,
problem-specific types) in the central `icfp` library crate.

The hope is that this split will allow everyone to prototype
asynchronously and with fewer merge conflicts, with the best code
being pulled out into the library and `submit` binary for submission.